requirements.txt
django==2.2.6
sqlparse==0.3.0
pytz==2019.3
django-mdeditor==0.1.16